refactor mysql concat_partition
