+++ src/concat/non_standard.rs
@@ -1,4 +1,4 @@
-#[cfg(any(feature = "postgresql", feature = "sqlite"))]
+#[cfg(any(feature = "postgresql", feature = "sqlite", feature = "mysql"))]
 use crate::{behavior::WithQuery, concat::concat_raw_before_after, fmt};
 
 #[cfg(any(feature = "postgresql", feature = "sqlite"))]
@@ -51,7 +51,7 @@ pub(crate) trait ConcatReturning<Clause: PartialEq> {
   }
 }
 
-#[cfg(any(feature = "postgresql", feature = "sqlite"))]
+#[cfg(any(feature = "postgresql", feature = "sqlite", feature = "mysql"))]
 pub(crate) trait ConcatWith<Clause: PartialEq> {
   fn concat_with(
     &self,
